requirements.txt
pandas
streamlit
PyPDF2
python-dotenv
langchain
langchain-community
openai
chromadb
tiktoken
# -e .